+++ retz-client/src/main/java/io/github/retz/web/Client.java
@@ -16,33 +16,42 @@
  */
 package io.github.retz.web;
 
-import java.io.IOException;
-import java.net.URI;
-import java.security.KeyManagementException;
-import java.security.NoSuchAlgorithmException;
-import java.util.Objects;
-import java.util.ResourceBundle;
-
-import javax.net.ssl.HostnameVerifier;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.SSLSocketFactory;
-import javax.net.ssl.TrustManager;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import feign.FeignException;
+import io.github.retz.auth.AuthHeader;
 import io.github.retz.auth.Authenticator;
+import io.github.retz.cli.TimestampHelper;
+import io.github.retz.misc.Pair;
+import io.github.retz.protocol.DownloadFileRequest;
 import io.github.retz.protocol.GetJobResponse;
 import io.github.retz.protocol.Response;
 import io.github.retz.protocol.ScheduleResponse;
 import io.github.retz.protocol.data.Application;
 import io.github.retz.protocol.data.Job;
 import io.github.retz.web.feign.Retz;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.net.ssl.HostnameVerifier;
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.SSLSocketFactory;
+import javax.net.ssl.TrustManager;
+import java.io.BufferedInputStream;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.net.HttpURLConnection;
+import java.net.URI;
+import java.net.URL;
+import java.security.KeyManagementException;
+import java.security.NoSuchAlgorithmException;
+import java.util.Objects;
+import java.util.ResourceBundle;
+
+import static java.nio.charset.StandardCharsets.UTF_8;
 
 public class Client implements AutoCloseable {
 
     public static final String VERSION_STRING;
+    public static final int MAX_BIN_SIZE = (int) DownloadFileRequest.MAX_FILE_SIZE;
 
     static final Logger LOG = LoggerFactory.getLogger(Client.class);
 
@@ -53,19 +62,23 @@ public class Client implements AutoCloseable {
 
     private final Retz retz;
     private boolean verboseLog = false;
+    private URI uri;
+    private Authenticator authenticator;
 
     protected Client(URI uri, Authenticator authenticator) {
         this(uri, authenticator, true);
     }
 
     protected Client(URI uri, Authenticator authenticator, boolean checkCert) {
+        this.uri = Objects.requireNonNull(uri);
+        this.authenticator = Objects.requireNonNull(authenticator);
         SSLSocketFactory socketFactory;
         HostnameVerifier hostnameVerifier;
         if (uri.getScheme().equals("https") && !checkCert) {
             LOG.warn("DANGER ZONE: TLS certificate check is disabled. Set 'retz.tls.insecure = false' at config file to supress this message.");
             try {
                 SSLContext sc = SSLContext.getInstance("SSL");
-                sc.init(null, new TrustManager[] { new WrongTrustManager() }, new java.security.SecureRandom());
+                sc.init(null, new TrustManager[]{new WrongTrustManager()}, new java.security.SecureRandom());
                 socketFactory = sc.getSocketFactory();
                 hostnameVerifier = new NoOpHostnameVerifier();
             } catch (NoSuchAlgorithmException e) {
@@ -81,12 +94,12 @@ public class Client implements AutoCloseable {
         System.setProperty("http.agent", Client.VERSION_STRING);
     }
 
-    public void setVerboseLog(boolean b) {
-        verboseLog = b;
-    }
-
     public static ClientBuilder newBuilder(URI uri) {
         return new ClientBuilder(uri);
+    }
+
+    public void setVerboseLog(boolean b) {
+        verboseLog = b;
     }
 
     @Override
@@ -124,6 +137,64 @@ public class Client implements AutoCloseable {
     public Response getFile(int id, String file, long offset, long length) throws IOException {
         return Retz.tryOrErrorResponse(
                 () -> retz.getFile(id, Objects.requireNonNull(file), offset, length));
+    }
+
+    public Pair<Integer, byte[]> getBinaryFile(int id, String file) throws IOException {
+        String date = TimestampHelper.now();
+        String resource = "/job/" + id + "/download?path=" + file;
+        AuthHeader header = authenticator.header("GET", "", date, resource);
+        URL url = new URL(uri.getScheme() + "://" + uri.getHost() + ":" + uri.getPort() + resource); // TODO url-encode!
+        LOG.info("Fetching {}", url);
+        HttpURLConnection conn;
+
+        conn = (HttpURLConnection) url.openConnection();
+        conn.setRequestMethod("GET");
+        conn.setRequestProperty("Accept", "application/octet-stream");
+        conn.setRequestProperty("Authorization", header.buildHeader());
+        conn.setRequestProperty("Date", date);
+        conn.setRequestProperty("Content-md5", "");
+        conn.setDoInput(true);
+        String s2s = authenticator.string2sign("GET", "", date, resource);
+        LOG.debug("Authorization: {} / S2S={}", header.buildHeader(), s2s);
+
+        if (conn.getResponseCode() != 200) {
+            if (conn.getResponseCode() < 200) {
+                throw new AssertionError(conn.getResponseMessage());
+            } else if (conn.getResponseCode() < 300) {
+                return new Pair<>(conn.getResponseCode(), "".getBytes(UTF_8)); // Mostly 204; success
+            } else if (conn.getResponseCode() < 400) {
+                return new Pair<>(conn.getResponseCode(), conn.getResponseMessage().getBytes(UTF_8));
+            } else if (conn.getResponseCode() == 404) {
+                throw new FileNotFoundException(url.toString());
+            } else {
+                return new Pair<>(conn.getResponseCode(), conn.getResponseMessage().getBytes(UTF_8));
+            }
+        }
+
+        int size = conn.getContentLength();
+        if (size < 0) {
+            throw new IOException("Illegal content length:" + size);
+        } else if (size == 0) {
+            // not bytes to save;
+            return new Pair<>(conn.getResponseCode(), new byte[0]);
+        } else if (size > MAX_BIN_SIZE) {
+            throw new IOException("Download file too large: " + size);
+        }
+        byte[] buffer = new byte[size];
+        try (BufferedInputStream in = new BufferedInputStream(conn.getInputStream())) {
+
+            int offset = 0;
+            while (offset < size) {
+                int read = in.read(buffer, offset, size - offset);
+                if (read < 0) {
+                    break;
+                }
+                offset += read;
+            }
+            return new Pair<>(conn.getResponseCode(), buffer);
+        } finally {
+            conn.disconnect();
+        }
     }
 
     public Response listFiles(int id, String path) throws IOException {